Fixing the default value for the TOneMapping curve and cleaning the code into the ToneMappingEffect file
