+++ libraries/render-utils/src/ToneMappingEffect.h
@@ -17,6 +17,8 @@
 
 #include <gpu/Resource.h>
 #include <gpu/Pipeline.h>
+#include <render/DrawTask.h>
+
 
 class RenderArgs;
 
@@ -50,7 +52,7 @@ private:
         float _exposure = 0.0f;
         float _twoPowExposure = 1.0f;
         glm::vec2 spareA;
-        int _toneCurve = Filmic;
+        int _toneCurve = Gamma22;
         glm::vec3 spareB;
 
         Parameters() {}
@@ -59,6 +61,31 @@ private:
     gpu::BufferView _parametersBuffer;
 
     void init();
+};
+
+class ToneMappingConfig : public render::Job::Config {
+    Q_OBJECT
+        Q_PROPERTY(bool enabled MEMBER enabled)
+        Q_PROPERTY(float exposure MEMBER exposure NOTIFY dirty);
+    Q_PROPERTY(int curve MEMBER curve NOTIFY dirty);
+public:
+    ToneMappingConfig() : render::Job::Config(true) {}
+
+    float exposure{ 0.0f };
+    int curve{ 1 };
+signals:
+    void dirty();
+};
+
+class ToneMappingDeferred {
+public:
+    using Config = ToneMappingConfig;
+    using JobModel = render::Job::Model<ToneMappingDeferred, Config>;
+
+    void configure(const Config& config);
+    void run(const render::SceneContextPointer& sceneContext, const render::RenderContextPointer& renderContext);
+
+    ToneMappingEffect _toneMappingEffect;
 };
 
 #endif // hifi_ToneMappingEffect_h
